Update app.js
allows for attacks made by a bullet currently doesnt interact with the hp of a entity, need to refactor so there is a initiliser packet (init) with all starting variables i.e max hp image sprites and such. a update packet which is on a constant loop that changes things like hp player position bullets shot and melle attacks. i also need to split almost all of this into the server folders.